Profile2->CompareResults for screen name, and added team colors for player profiles, leak canary added to detect leaks too
