+++ app/build.gradle.kts
@@ -58,7 +58,7 @@ dependencies {
     implementation("androidx.core:core-ktx:1.12.0")
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.7.0")
     implementation("androidx.activity:activity-compose:1.8.2")
-    implementation(platform("androidx.compose:compose-bom:2023.08.00"))
+    implementation(platform("androidx.compose:compose-bom:2024.03.00"))
     implementation("androidx.compose.ui:ui")
     implementation("androidx.compose.ui:ui-graphics")
     implementation("androidx.compose.ui:ui-tooling-preview")
@@ -72,20 +72,20 @@ dependencies {
     //For loading images
     implementation("io.coil-kt:coil-compose:2.6.0")
     //For extra icons
-    implementation("androidx.compose.material:material-icons-extended:1.6.3")
+    implementation("androidx.compose.material:material-icons-extended:1.6.4")
     implementation("com.google.android.gms:play-services-dtdi:16.0.0-beta01")
     //For time operations
     implementation("com.jakewharton.threetenabp:threetenabp:1.4.6")
     //For time picker on games screen, we needed to change repositories in settings.gradle.kts
-    implementation ("com.github.Rafsanjani:datepickertimeline:0.7.1")
-
+    implementation("com.github.Rafsanjani:datepickertimeline:0.7.1")
+    //For detecting leaks in notification of emulator
+    debugImplementation("com.squareup.leakcanary:leakcanary-android:2.13")
 
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-    androidTestImplementation(platform("androidx.compose:compose-bom:2023.08.00"))
     androidTestImplementation("androidx.compose.ui:ui-test-junit4")
-    androidTestImplementation(platform("androidx.compose:compose-bom:2023.08.00"))
+    androidTestImplementation(platform("androidx.compose:compose-bom:2024.03.00"))
     debugImplementation("androidx.compose.ui:ui-tooling")
     debugImplementation("androidx.compose.ui:ui-test-manifest")
 }
