+++ app/src/main/java/com/example/jetpacktest/screens/CompareScreen.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
@@ -19,8 +18,6 @@ import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TextFieldDefaults
@@ -43,8 +40,9 @@ import com.example.jetpacktest.DatabaseHandler
 fun CompareScreen(
     navigateToPlayerProfile: (String) -> Unit,
     navController: NavHostController,
-    navigateToProfile2: (List<String>) -> Unit
+    navigateToCompareResults: (List<String>) -> Unit
 ) {
+
     var searchText by remember { mutableStateOf("") }
     val focusManager = LocalFocusManager.current
     var searchResults by remember { mutableStateOf<List<String>>(emptyList()) }
@@ -52,44 +50,61 @@ fun CompareScreen(
     var selectedSearchType by remember { mutableStateOf("Player") }
     var showAlertDialog by remember { mutableStateOf(false) }
     var showSelectionDialog by remember { mutableStateOf(false) }
-
     val databaseHandler = DatabaseHandler()
 
-    Surface(
-        modifier = Modifier.fillMaxSize(),
-        color = MaterialTheme.colorScheme.background
-    ) {
-        Column(modifier = Modifier.padding(10.dp)) {
-            RadioButtonsDisplay(
-                selectedSearchType = selectedSearchType,
-                onSearchTypeSelected = { searchType ->
-                    selectedSearchType = searchType
-                },
-                onSearchTextChange = { newText ->
-                    searchText = newText
-                    if (selectedSearchType == "Team") {
-                        handleTeamSearch(searchText) { newResults ->
-                            searchResults = newResults.take(8)
-                        }
-                    } else {
-                        databaseHandler.executePlayerSearchResults(searchText) { newResults ->
-                            searchResults = newResults.take(8)
-                        }
+    Column(modifier = Modifier.padding(10.dp)) {
+        RadioButtonsDisplay(
+            selectedSearchType = selectedSearchType,
+            onSearchTypeSelected = { searchType ->
+                selectedSearchType = searchType
+            },
+            onSearchTextChange = { newText ->
+                searchText = newText
+                if (selectedSearchType == "Team") {
+                    handleTeamSearch(searchText) { newResults ->
+                        searchResults = newResults.take(8)
                     }
-                },
-                onClearSearchResults = {
-                    searchResults = emptyList()
+                } else {
+                    databaseHandler.executePlayerSearchResults(searchText) { newResults ->
+                        searchResults = newResults.take(8)
+                    }
                 }
-            )
-            Spacer(modifier = Modifier.height(6.dp))
-            TextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .clip(RoundedCornerShape(8.dp)),
-                value = searchText,
-                singleLine = true,
-                onValueChange = {
-                    searchText = it
+            },
+            onClearSearchResults = {
+                searchResults = emptyList()
+            }
+        )
+        Spacer(modifier = Modifier.height(6.dp))
+        TextField(
+            modifier = Modifier
+                .fillMaxWidth()
+                .clip(RoundedCornerShape(8.dp)),
+            value = searchText,
+            singleLine = true,
+            onValueChange = {
+                searchText = it
+                if (selectedSearchType == "Player") {
+                    databaseHandler.executePlayerSearchResults(searchText) { newResults ->
+                        searchResults = newResults.take(5)
+                    }
+                } else {
+                    handleTeamSearch(searchText) { newResults ->
+                        searchResults = newResults.take(5)
+                    }
+                }
+            },
+            label = {
+                Text(if (selectedSearchType == "Player") "Search Players ..."
+                else "Search Teams...")
+            },
+            leadingIcon = {
+                Icon(
+                    imageVector = Icons.Default.Search,
+                    contentDescription = "Search Icon"
+                )
+            },
+            keyboardActions = KeyboardActions(
+                onDone = {
                     if (selectedSearchType == "Player") {
                         databaseHandler.executePlayerSearchResults(searchText) { newResults ->
                             searchResults = newResults.take(5)
@@ -99,86 +114,63 @@ fun CompareScreen(
                             searchResults = newResults.take(5)
                         }
                     }
-                },
-
-                label = {
-                    Text(if (selectedSearchType == "Player") "Search Players ..."
-                    else "Search Teams...")
-                },
-                leadingIcon = {
-                    Icon(
-                        imageVector = Icons.Default.Search,
-                        contentDescription = "Search Icon"
-                    )
-                },
-                keyboardActions = KeyboardActions(
-                    onDone = {
-                        if (selectedSearchType == "Player") {
-                            databaseHandler.executePlayerSearchResults(searchText) { newResults ->
-                                searchResults = newResults.take(5)
-                            }
-                        } else {
-                            handleTeamSearch(searchText) { newResults ->
-                                searchResults = newResults.take(5)
+                    focusManager.clearFocus()
+                }
+            ),
+            colors = TextFieldDefaults.colors(
+                cursorColor = Color.Blue,
+                focusedIndicatorColor = Color.Transparent,
+                unfocusedIndicatorColor = Color.Transparent
+            )
+        )
+        Spacer(modifier = Modifier.height(16.dp))
+        LazyColumn(modifier = Modifier.weight(1f)) {
+            items(searchResults) { itemName ->
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(vertical = 8.dp)
+                        .clickable {
+                            if (selectedPlayers.size < 2) {
+                                selectedPlayers = selectedPlayers + itemName
+                            } else {
+                                // Display warning or exception for selecting more than two players
+                                // For now, just print a warning
+                                showAlertDialog = true
+                                println("Cannot select more than two players")
                             }
                         }
-                        focusManager.clearFocus()
-                    }
-                ),
-                colors = TextFieldDefaults.colors(
-                    cursorColor = Color.Blue,
-                    focusedIndicatorColor = Color.Transparent,
-                    unfocusedIndicatorColor = Color.Transparent
-                )
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-            LazyColumn(modifier = Modifier.weight(1f)) {
-                items(searchResults) { itemName ->
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(vertical = 8.dp)
-                            .clickable {
-                                if (selectedPlayers.size < 2) {
-                                    selectedPlayers = selectedPlayers + itemName
-                                } else {
-                                    // Display warning or exception for selecting more than two players
-                                    // For now, just print a warning
-                                    showAlertDialog = true
-                                    println("Cannot select more than two players")
-                                }
-                            }
-                    ) {
-                        Text(text = itemName)
-                    }
+                ) {
+                    Text(text = itemName)
                 }
             }
-            if (selectedPlayers.isNotEmpty()) {
-                Text("Selected Players:")
-                LazyColumn {
-                    items(selectedPlayers) { playerName ->
-                        Row(verticalAlignment = Alignment.CenterVertically) {
+        }
+        if (selectedPlayers.isNotEmpty()) {
+            Text("Selected Players:")
+            LazyColumn {
+                items(selectedPlayers) { playerName ->
+                    Row(verticalAlignment = Alignment.CenterVertically) {
                         Text(text = playerName)
                         Spacer(modifier = Modifier.weight(1f))
                         // Red X button to remove player
-                        Box(modifier = Modifier.clickable {
-                            selectedPlayers = selectedPlayers.filter { it != playerName }
-                        }) {
+                        Box(modifier = Modifier
+                            .clickable { selectedPlayers = selectedPlayers
+                                .filter { it != playerName } } ) {
                             Icon(
                                 imageVector = Icons.Default.Clear,
                                 contentDescription = "Remove Player",
                                 tint = Color.Red
-                        )}
-                }
+                            )
+                        }
                     }
+                }
             }
             Button(
                 onClick = {
                     if (selectedPlayers.size < 2) {
                         showSelectionDialog = true
                     } else {
-                        // Navigate to profile2 screen with selected players
-                        navController.navigate("profile2Screen")
+                        navController.navigate("compareresultsscreen")
                     }
                 },
                 modifier = Modifier.padding(vertical = 16.dp)
@@ -215,5 +207,4 @@ fun CompareScreen(
             }
         )
     }
-
-}}
+}
