+++ app/src/main/res/drawable/xml_celtics.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="214.48dp"
-    android:height="214.48dp"
+    android:width="200.48dp"
+    android:height="200.48dp"
     android:viewportWidth="214.48"
     android:viewportHeight="214.48">
   <path
